Changed SNR to float - now DG8245V correctly calculates its SNR values

+++ internal/modem/model.go
@@ -1,21 +1,21 @@
 package modem
 
 type Stats struct {
-	Uptime      int64 `json:"uptime"`
-	Status      bool  `json:"status"`
-	VoipStatus  bool  `json:"voip_status"`
-	CurrentUp   int   `json:"current_up"`
-	CurrentDown int   `json:"current_down"`
-	MaxUp       int   `json:"max_up"`
-	MaxDown     int   `json:"max_down"`
-	DataUp      int64 `json:"data_up"`
-	DataDown    int64 `json:"data_down"`
-	FECUp       int   `json:"fec_up"`
-	FECDown     int   `json:"fec_down"`
-	CRCUp       int   `json:"crc_up"`
-	CRCDown     int   `json:"crc_down"`
-	SNRUp       int64 `json:"snr_up"`
-	SNRDown     int64 `json:"snr_down"`
+	Uptime      int64   `json:"uptime"`
+	Status      bool    `json:"status"`
+	VoipStatus  bool    `json:"voip_status"`
+	CurrentUp   int     `json:"current_up"`
+	CurrentDown int     `json:"current_down"`
+	MaxUp       int     `json:"max_up"`
+	MaxDown     int     `json:"max_down"`
+	DataUp      int64   `json:"data_up"`
+	DataDown    int64   `json:"data_down"`
+	FECUp       int     `json:"fec_up"`
+	FECDown     int     `json:"fec_down"`
+	CRCUp       int     `json:"crc_up"`
+	CRCDown     int     `json:"crc_down"`
+	SNRUp       float64 `json:"snr_up"`
+	SNRDown     float64 `json:"snr_down"`
 }
 
 type Statable interface {
